% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{scale_per_charge}
\alias{scale_per_charge}
\title{Z scaling}
\usage{
scale_per_charge(data, charge_col, cols_to_scale)
}
\arguments{
\item{data}{frame}

\item{charge_col}{name of col}

\item{cols_to_scale}{}
}
\value{
A dataframe with additional z scaled coloumns
}
\description{
Z scaling
}
\examples{
jojo

}
